+++ docs/README.md
@@ -1,4 +1,4 @@
-# Elena User Guide
+# elena.Elena User Guide
 
 // Update the title above to match the actual product name
