+++ src/main/java/elena/AddTodoCommand.java
@@ -1,3 +1,5 @@
+package elena;
+
 public class AddTodoCommand implements Command {
     private final String input;
