+++ src/main/java/elena/Command.java
@@ -1,3 +1,5 @@
+package elena;
+
 public interface Command {
     void execute(TaskList tasks, Ui ui, Storage storage) throws ElenaException;
     boolean isExit();
