+++ src/main/java/elena/ElenaException.java
@@ -1,3 +1,5 @@
+package elena;
+
 public class ElenaException extends Exception {
     public ElenaException(String message) { super(message); }
 
@@ -5,10 +7,10 @@ public class ElenaException extends Exception {
         return new ElenaException("Your todo cannot be empty! Please type something after 'todo'.");
     }
     public static ElenaException emptyDeadline() {
-        return new ElenaException("Deadline missing description or /by time! Use: deadline <desc> /by <time>");
+        return new ElenaException("elena.Deadline missing description or /by time! Use: deadline <desc> /by <time>");
     }
     public static ElenaException emptyEvent() {
-        return new ElenaException("Event missing description or time! Use: event <desc> /from <start> /to <end>");
+        return new ElenaException("elena.Event missing description or time! Use: event <desc> /from <start> /to <end>");
     }
     public static ElenaException invalidCommand(String command) {
         return new ElenaException("Oops! I don't understand '" + command + "'. Try a valid command.");
@@ -17,6 +19,6 @@ public class ElenaException extends Exception {
         return new ElenaException("The task number you provided is invalid. Make sure it exists in your list.");
     }
     public static ElenaException nonIntegerTaskNumber() {
-        return new ElenaException("Task number must be a whole number. Example: mark 2");
+        return new ElenaException("elena.Task number must be a whole number. Example: mark 2");
     }
 }
