+++ src/main/java/elena/Event.java
@@ -1,3 +1,5 @@
+package elena;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
