+++ src/main/java/elena/Parser.java
@@ -1,3 +1,5 @@
+package elena;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
