+++ src/main/java/elena/Storage.java
@@ -1,3 +1,5 @@
+package elena;
+
 import java.io.*;
 import java.nio.file.*;
 import java.time.LocalDateTime;
