+++ src/main/java/elena/Task.java
@@ -1,3 +1,5 @@
+package elena;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
