+++ src/main/java/elena/Todo.java
@@ -1,3 +1,5 @@
+package elena;
+
 public class Todo extends Task {
     public Todo(String description) { super(description, TaskType.TODO); }
