+++ src/main/java/elena/Ui.java
@@ -1,3 +1,5 @@
+package elena;
+
 import java.util.Scanner;
 
 public class Ui {
@@ -9,7 +11,7 @@ public class Ui {
 
     public void showWelcome() {
         showLine();
-        System.out.println(" Hello! I'm Elena 🤖");
+        System.out.println(" Hello! I'm elena.Elena 🤖");
         System.out.println(" What can I do for you?");
         showLine();
     }
