+++ src/main/java/elena/UnmarkCommand.java
@@ -1,3 +1,5 @@
+package elena;
+
 public class UnmarkCommand implements Command {
     private final String input;
