+++ text-ui-test/EXPECTED.TXT
@@ -1,5 +1,5 @@
 ____________________________________________________________
- Hello! I'm Elena 🤖
+ Hello! I'm elena.Elena 🤖
  What can I do for you?
 ____________________________________________________________
 ____________________________________________________________
